no_std for cairo-lang-casm and cairo-lang-utils
add ensure_no_std crate

move Felt252 as_short_string fn to cairo-lang-utils

remove unused deps from cario-lang-runner

add crate cairo-lang-vm-utils

add crate cairo-lang-casm-contract-class

upgrade cairo-rs dependencies to 0.6.0

build: bump indexmap to 2.0.0

format: ran cargo fmt
