+++ Cargo.toml
@@ -5,6 +5,7 @@ overflow-checks = true     # Enable integer overflow checks.
 
 members = [
     "crates/cairo-lang-casm",
+    "crates/cairo-lang-casm-contract-class",
     "crates/cairo-lang-compiler",
     "crates/cairo-lang-debug",
     "crates/cairo-lang-defs",
@@ -31,6 +32,7 @@ members = [
     "crates/cairo-lang-syntax-codegen",
     "crates/cairo-lang-test-runner",
     "crates/cairo-lang-utils",
+    "crates/cairo-lang-vm-utils",
     "crates/bin/cairo-language-server",
     "crates/bin/cairo-compile",
     "crates/bin/cairo-format",
@@ -43,6 +45,8 @@ members = [
     "tests",
 ]
 
+exclude = ["ensure_no_std"]
+
 [workspace.package]
 version = "2.0.0-rc5"
 edition = "2021"
@@ -52,14 +56,14 @@ license-file = "LICENSE"
 
 [workspace.dependencies]
 anyhow = "1.0.66"
-ark-ff = "0.4.0-alpha.7"
+ark-std = { version = "0.3.0", default-features = false }
+ark-ff = { version = "0.4.0-alpha.7", default-features = false }
 ark-secp256k1 = "0.4.0"
 ark-secp256r1 = "0.4.0"
-ark-std = "0.3.0"
 assert_matches = "1.5"
 bimap = "0.6.2"
-cairo-felt = "0.8.1"
-cairo-vm = "0.8.1"
+cairo-felt = {version = "0.8.1", default-features = false }
+cairo-vm = {  version = "0.8.1", default-features = false }
 clap = { version = "4.0", features = ["derive"] }
 colored = "2"
 const-fnv1a-hash = "1.1.0"
@@ -69,20 +73,25 @@ diffy = "0.3.0"
 env_logger = "0.9.3"
 genco = "0.17.0"
 good_lp = { version = "1.3.2", features = ["minilp"], default-features = false }
+hashbrown = { version = "0.14.0", features = ["serde"] }
 id-arena = "2.2.1"
 ignore = "0.4.20"
-indexmap = { version = "1.9.1", features = ["serde"] }
-indoc = "2.0.1"
-itertools = "0.10.3"
+indexmap = {version = "2.0.0", features = [
+    "serde",
+], default-features = false }
+indoc = { version = "2.0.1", default-features = false }
+itertools = { version = "0.10.5", default-features = false, features = [
+    "use_alloc",
+] }
 keccak = "0.1.3"
 lalrpop-util = { version = "0.19.9", features = ["lexer"] }
 log = "0.4"
 lsp = { version = "0.93", package = "lsp-types" }
-num-bigint = "0.4"
-num-integer = "0.1"
-num-traits = "0.2"
+num-bigint = { version = "0.4", default-features = false, features = ["serde"] }
+num-integer = { version = "0.1", default-features = false }
+num-traits = { version = "0.2", default-features = false }
 once_cell = "1.17.1"
-parity-scale-codec = "3.5.0"
+parity-scale-codec = { version = "3.5.0", default-features = false }
 parity-scale-codec-derive = "3.1.4"
 path-clean = "0.1.0"
 pretty_assertions = "1.2.1"
@@ -93,7 +102,10 @@ rstest = "0.16.0"
 salsa = "0.16.1"
 scarb-metadata = "1.3.0"
 schemars = { version = "0.8.12", features = ["preserve_order"] }
-serde = { version = "1.0.130", features = ["derive"] }
+serde = { version = "1.0.130", default-features = false, features = [
+    "derive",
+    "alloc",
+] }
 serde_json = "1.0"
 sha3 = "0.10.6"
 smol_str = { version = "0.2.0", features = ["serde"] }
