+++ crates/bin/starknet-sierra-compile/Cargo.toml
@@ -12,6 +12,7 @@ clap.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 
-cairo-lang-sierra = { path = "../../cairo-lang-sierra", version = "2.0.0-rc5" }
-cairo-lang-starknet = { path = "../../cairo-lang-starknet", version = "2.0.0-rc5" }
-cairo-lang-utils = { path = "../../cairo-lang-utils", version = "2.0.0-rc5" }
+cairo-lang-sierra = { path = "../../cairo-lang-sierra" }
+cairo-lang-starknet = { path = "../../cairo-lang-starknet" }
+cairo-lang-utils = { path = "../../cairo-lang-utils" }
+cairo-lang-casm-contract-class = { path = "../../cairo-lang-casm-contract-class" }
