+++ crates/cairo-lang-casm/src/hints/mod.rs
@@ -1,9 +1,8 @@
-use std::fmt::{Display, Formatter};
+use core::fmt::{Display, Formatter};
 
 use cairo_lang_utils::bigint::BigIntAsHex;
 use indoc::formatdoc;
 use parity_scale_codec_derive::{Decode, Encode};
-use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
 use crate::operand::{CellRef, DerefOrImmediate, ResOperand};
@@ -14,7 +13,8 @@ mod test;
 // Represents a cairo hint.
 // Note: Hint encoding should be backwards-compatible. This is an API guarantee.
 // For example, new variants should have new `index`.
-#[derive(Serialize, Deserialize, Debug, Eq, PartialEq, Clone, Encode, Decode, JsonSchema)]
+#[derive(Serialize, Deserialize, Debug, Eq, PartialEq, Clone, Encode, Decode)]
+#[cfg_attr(feature = "schemars", derive(schemars::JsonSchema))]
 #[serde(untagged)]
 pub enum Hint {
     #[codec(index = 0)]
@@ -56,7 +56,8 @@ impl PythonicHint for Hint {
 }
 
 /// Represents a hint that triggers a system call.
-#[derive(Serialize, Deserialize, Debug, Eq, PartialEq, Clone, Encode, Decode, JsonSchema)]
+#[derive(Serialize, Deserialize, Debug, Eq, PartialEq, Clone, Encode, Decode)]
+#[cfg_attr(feature = "schemars", derive(schemars::JsonSchema))]
 pub enum StarknetHint {
     #[codec(index = 0)]
     SystemCall { system: ResOperand },
@@ -71,7 +72,8 @@ pub enum StarknetHint {
 }
 
 // Represents a cairo core hint.
-#[derive(Serialize, Deserialize, Debug, Eq, PartialEq, Clone, Encode, Decode, JsonSchema)]
+#[derive(Serialize, Deserialize, Debug, Eq, PartialEq, Clone, Encode, Decode)]
+#[cfg_attr(feature = "schemars", derive(schemars::JsonSchema))]
 #[serde(untagged)]
 pub enum CoreHintBase {
     #[codec(index = 0)]
@@ -91,7 +93,8 @@ impl From<DeprecatedHint> for CoreHintBase {
     }
 }
 
-#[derive(Serialize, Deserialize, Debug, Eq, PartialEq, Clone, Encode, Decode, JsonSchema)]
+#[derive(Serialize, Deserialize, Debug, Eq, PartialEq, Clone, Encode, Decode)]
+#[cfg_attr(feature = "schemars", derive(schemars::JsonSchema))]
 pub enum CoreHint {
     #[codec(index = 0)]
     AllocSegment { dst: CellRef },
@@ -228,7 +231,8 @@ pub enum CoreHint {
 
 /// Represents a deprecated hint which is kept for backward compatibility of previously deployed
 /// contracts.
-#[derive(Serialize, Deserialize, Debug, Eq, PartialEq, Clone, Encode, Decode, JsonSchema)]
+#[derive(Serialize, Deserialize, Debug, Eq, PartialEq, Clone, Encode, Decode)]
+#[cfg_attr(feature = "schemars", derive(schemars::JsonSchema))]
 pub enum DeprecatedHint {
     /// Asserts that the current access indices list is empty (after the loop).
     #[codec(index = 0)]
@@ -257,7 +261,7 @@ pub enum DeprecatedHint {
 
 struct DerefOrImmediateFormatter<'a>(&'a DerefOrImmediate);
 impl<'a> Display for DerefOrImmediateFormatter<'a> {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+    fn fmt(&self, f: &mut Formatter<'_>) -> core::fmt::Result {
         match self.0 {
             DerefOrImmediate::Deref(d) => write!(f, "memory{d}"),
             DerefOrImmediate::Immediate(i) => write!(f, "{}", i.value),
@@ -267,7 +271,7 @@ impl<'a> Display for DerefOrImmediateFormatter<'a> {
 
 struct ResOperandFormatter<'a>(&'a ResOperand);
 impl<'a> Display for ResOperandFormatter<'a> {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+    fn fmt(&self, f: &mut Formatter<'_>) -> core::fmt::Result {
         match self.0 {
             ResOperand::Deref(d) => write!(f, "memory{d}"),
             ResOperand::DoubleDeref(d, i) => write!(f, "memory[memory{d} + {i}]"),
