+++ crates/cairo-lang-casm/src/inline.rs
@@ -1,3 +1,6 @@
+#[cfg(not(feature = "std"))]
+use alloc::vec::Vec;
+
 use crate::hints::Hint;
 use crate::instructions::Instruction;
 
@@ -215,7 +218,7 @@ macro_rules! casm_extend {
 #[macro_export]
 macro_rules! append_instruction {
     ($ctx:ident, $body:ident $(,$ap:ident++)?) => {
-        let current_hints = std::mem::take(&mut $ctx.current_hints);
+        let current_hints = core::mem::take(&mut $ctx.current_hints);
         let instr = $crate::instructions::Instruction {
             body: $body,
             inc_ap: $crate::is_inc_ap!($($ap++)?),
