+++ crates/cairo-lang-casm/src/operand_test.rs
@@ -1,3 +1,6 @@
+#[cfg(not(feature = "std"))]
+use alloc::string::ToString;
+
 use test_log::test;
 
 use super::{BinOpOperand, DerefOrImmediate, Operation};
