+++ crates/cairo-lang-runner/src/lib.rs
@@ -39,7 +39,6 @@ use num_traits::ToPrimitive;
 use thiserror::Error;
 
 pub mod casm_run;
-pub mod short_string;
 
 #[derive(Debug, Error)]
 pub enum RunnerError {
