+++ crates/cairo-lang-starknet/Cargo.toml
@@ -9,22 +9,23 @@ description = "Starknet capabilities and utilities on top of Cairo."
 [dependencies]
 anyhow.workspace = true
 cairo-felt.workspace = true
-cairo-lang-casm = { path = "../cairo-lang-casm", version = "2.0.0-rc5" }
-cairo-lang-compiler = { path = "../cairo-lang-compiler", version = "2.0.0-rc5" }
-cairo-lang-defs = { path = "../cairo-lang-defs", version = "2.0.0-rc5" }
-cairo-lang-diagnostics = { path = "../cairo-lang-diagnostics", version = "2.0.0-rc5" }
-cairo-lang-filesystem = { path = "../cairo-lang-filesystem", version = "2.0.0-rc5" }
-cairo-lang-lowering = { path = "../cairo-lang-lowering", version = "2.0.0-rc5" }
-cairo-lang-parser = { path = "../cairo-lang-parser", version = "2.0.0-rc5" }
-cairo-lang-plugins = { path = "../cairo-lang-plugins", version = "2.0.0-rc5" }
-cairo-lang-semantic = { path = "../cairo-lang-semantic", version = "2.0.0-rc5" }
-cairo-lang-sierra = { path = "../cairo-lang-sierra", version = "2.0.0-rc5" }
-cairo-lang-sierra-ap-change = { path = "../cairo-lang-sierra-ap-change", version = "2.0.0-rc5" }
-cairo-lang-sierra-gas = { path = "../cairo-lang-sierra-gas", version = "2.0.0-rc5" }
-cairo-lang-sierra-generator = { path = "../cairo-lang-sierra-generator", version = "2.0.0-rc5" }
-cairo-lang-sierra-to-casm = { path = "../cairo-lang-sierra-to-casm", version = "2.0.0-rc5" }
-cairo-lang-syntax = { path = "../cairo-lang-syntax", version = "2.0.0-rc5" }
-cairo-lang-utils = { path = "../cairo-lang-utils", version = "2.0.0-rc5" }
+cairo-lang-casm = { path = "../cairo-lang-casm" }
+cairo-lang-compiler = { path = "../cairo-lang-compiler" }
+cairo-lang-defs = { path = "../cairo-lang-defs" }
+cairo-lang-diagnostics = { path = "../cairo-lang-diagnostics" }
+cairo-lang-filesystem = { path = "../cairo-lang-filesystem" }
+cairo-lang-lowering = { path = "../cairo-lang-lowering" }
+cairo-lang-parser = { path = "../cairo-lang-parser" }
+cairo-lang-plugins = { path = "../cairo-lang-plugins" }
+cairo-lang-semantic = { path = "../cairo-lang-semantic" }
+cairo-lang-sierra = { path = "../cairo-lang-sierra" }
+cairo-lang-sierra-ap-change = { path = "../cairo-lang-sierra-ap-change" }
+cairo-lang-sierra-gas = { path = "../cairo-lang-sierra-gas" }
+cairo-lang-sierra-generator = { path = "../cairo-lang-sierra-generator" }
+cairo-lang-sierra-to-casm = { path = "../cairo-lang-sierra-to-casm" }
+cairo-lang-syntax = { path = "../cairo-lang-syntax" }
+cairo-lang-utils = { path = "../cairo-lang-utils" }
+cairo-lang-casm-contract-class = { path = "../cairo-lang-casm-contract-class" }
 convert_case.workspace = true
 genco.workspace = true
 indoc.workspace = true
