+++ crates/cairo-lang-utils/Cargo.toml
@@ -13,11 +13,15 @@ num-bigint.workspace = true
 num-integer.workspace = true
 num-traits.workspace = true
 parity-scale-codec.workspace = true
-schemars = { workspace = true, features = ["preserve_order"] }
 serde.workspace = true
+cairo-felt = { workspace = true, features = ["alloc"] }
+hashbrown.workspace = true
 
-# Optional
+# Optional features
+schemars = { workspace = true, optional = true }
 env_logger = { workspace = true, optional = true }
+
+# Optional dependencies
 time = { workspace = true, optional = true }
 log = { workspace = true, optional = true }
 
@@ -28,5 +32,15 @@ test-log.workspace = true
 env_logger.workspace = true
 
 [features]
+default = ["std"]
+std = [
+    "indexmap/std",
+    "itertools/use_std",
+    "num-bigint/std",
+    "num-integer/std",
+    "num-traits/std",
+    "serde/std",
+    "cairo-felt/std",
+]
+env_logger = ["std", "dep:env_logger", "dep:time", "dep:log"]
 testing = []
-env_logger = ["dep:env_logger", "dep:time", "dep:log"]
