+++ crates/cairo-lang-utils/src/graph_algos/scc_graph_node.rs
@@ -1,3 +1,6 @@
+#[cfg(not(feature = "std"))]
+use alloc::vec::Vec;
+
 use super::graph_node::GraphNode;
 use super::strongly_connected_components::ComputeScc;
