+++ crates/cairo-lang-utils/src/graph_algos/strongly_connected_components_test.rs
@@ -1,5 +1,10 @@
+#[cfg(not(feature = "std"))]
+use alloc::{vec, vec::Vec};
+#[cfg(feature = "std")]
 use std::collections::HashSet;
 
+#[cfg(not(feature = "std"))]
+use hashbrown::HashSet;
 use itertools::chain;
 use test_case::test_case;
 use test_log::test;
