+++ ensure_no_std/src/main.rs
@@ -1,0 +1,21 @@
+#![no_std]
+#![no_main]
+
+use core::panic::PanicInfo;
+
+/// This function is called on panic.
+#[panic_handler]
+fn panic(_info: &PanicInfo) -> ! {
+    loop {}
+}
+
+#[no_mangle]
+pub extern "C" fn _start() -> ! {
+    loop {}
+}
+
+#[global_allocator]
+static ALLOC: wee_alloc::WeeAlloc = wee_alloc::WeeAlloc::INIT;
+
+#[allow(unused_imports)]
+use {cairo_lang_casm, cairo_lang_casm_contract_class, cairo_lang_utils, cairo_lang_vm_utils};
